+++ Sources/CodexBar/StatusItemController+CountdownRefresh.swift
@@ -267,6 +267,9 @@ extension StatusItemController {
     }
 
     func handleMenuBarTimeEnvironmentChange() {
+        if self.store.codexRemoteCosts.enabled {
+            self.store.codexRemoteCosts.reconcile(self.store.codexRemoteCostContext())
+        }
         self.updateIcons()
     }